add authentication

+++ source/PortfolioTracker/PortfolioTracker/Controllers/PortfolioController.cs
@@ -6,6 +6,9 @@ using PortfolioTracker.Data.Dtos.Portfolios;
 using PortfolioTracker.Data.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using PortfolioTracker.Data;
+using Microsoft.AspNetCore.Authorization;
+using PortfolioTracker.Data.Dtos.Auth;
+using PortfolioTracker.Auth.Model;
 
 namespace PortfolioTracker.Controllers
 {
@@ -15,21 +18,26 @@ namespace PortfolioTracker.Controllers
     {
         private readonly IPortfolioRepository _portfolioRepository;
         private readonly IMapper _mapper;
+        private readonly IAuthorizationService _authorizationService;
 
-        public PortfolioController(IPortfolioRepository portfolioRepository, IMapper mapper)
+        public PortfolioController(IPortfolioRepository portfolioRepository, IMapper mapper, IAuthorizationService authorizationService)
         {
             _portfolioRepository = portfolioRepository;
             _mapper = mapper;
+            _authorizationService = authorizationService;
         }
 
         [HttpGet]
+        [Authorize(Roles = PortfolioUserRoles.SimpleUser)]
         public async Task<IEnumerable<PortfolioDto>> GetAllAsync()
         {
+            var userClaims = User;
             var portfolio = await _portfolioRepository.GetAsync();
             return portfolio.Select(portfolio => _mapper.Map<PortfolioDto>(portfolio));
         }
 
         [HttpGet("{portfolioId}")]
+        [Authorize(Roles = PortfolioUserRoles.SimpleUser)]
         public async Task<ActionResult<PortfolioDto>> GetAsync(int portfolioId)
         {
             var portfolio = await _portfolioRepository.GetAsync(portfolioId);
@@ -49,13 +57,18 @@ namespace PortfolioTracker.Controllers
         }
 
         [HttpPut("{portfolioId}")]
-        public async Task<ActionResult<PortfolioDto>> PostAsync(int portfolioId, UpdatePortfolioDto portfolioDto)
+        [Authorize(Roles = PortfolioUserRoles.SimpleUser)]
+        public async Task<ActionResult<PortfolioDto>> PutAsync(int portfolioId, UpdatePortfolioDto portfolioDto)
         {
             var oldPortfolio = await _portfolioRepository.GetAsync(portfolioId);
             if (oldPortfolio == null)
                 return NotFound();
 
-            //oldPost.Body = postDto.Body;
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, oldPortfolio, PolicyNames.SameUser);
+
+            if (!authorizationResult.Succeeded)
+                return Forbid();
+
             _mapper.Map(portfolioDto, oldPortfolio);
 
             await _portfolioRepository.UpdateAsync(oldPortfolio);
@@ -64,11 +77,17 @@ namespace PortfolioTracker.Controllers
         }
 
         [HttpDelete("{portfolioId}")]
+        [Authorize(Roles = PortfolioUserRoles.SimpleUser)]
         public async Task<ActionResult> DeleteAsync(int portfolioId)
         {
             var portfolio = await _portfolioRepository.GetAsync(portfolioId);
             if (portfolio == null)
                 return NotFound();
+
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, portfolio, PolicyNames.SameUser);
+
+            if (!authorizationResult.Succeeded)
+                return Forbid();
 
             await _portfolioRepository.DeleteAsync(portfolio);
